requirements.txt
astroid==2.4.2
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
github-traffic-stats2==20.7.9
githubpy==1.1.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauth2client==4.1.3
pickleDB==0.9.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pylint==2.5.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
simplejson==3.17.0
six==1.15.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
